fix(object): move array encoding from datachunk to its own class
